Separate JVM implementation to separate class
